changed file structure and made enum proto alphabetical
