+++ hologram_dashboard/fetcher/proto/enums.proto
@@ -15,41 +15,41 @@
 */
 
 syntax = "proto2";
-package wireless_android_play_analytics;
+package wireless_android_play_analytics,
 
 enum DataType {
     UNKNOWN_DATA_TYPE = 0;
-    APP_COUNTRY_PUBLISH_TIME = 1;
-    SPAM = 2;
-    QUERY_CATEGORY_SOURCE = 3;
-    APPS_DAILY_DATA_SOURCE = 4;
-    UNIFIED_USER_DATA_SOURCE = 5;
-    TEST_CODE = 6;
-    DEVICE_INTEGRITY_SOURCE = 7; 
-    SHIELD_ID = 8;
-    SERVER_EXPERIMENT = 9; 
-    APPS_TITLE_TIER = 10;
-    APP_TYPE_SOURCE = 11;
-    ANDROID_SPAM_ABUSE = 12;
-    USER_EXPERIMENT_ID_AND_TEST_CODE = 13;
-    USER_ATTRIBUTE = 14;
-    PRIME_MERIDIAN= 15;
-    PLAY_COUNTRY = 16;
+    ANDROID_SPAM_ABUSE = 1;
+    APPS_DAILY_DATA_SOURCE = 2;
+    APPS_TITLE_TIER = 3;
+    APP_COUNTRY_PUBLISH_TIME = 4;
+    APP_TYPE_SOURCE = 5;
+    DEVICE_INTEGRITY_SOURCE = 6;
+    PLAY_COUNTRY = 7;
+    PRIME_MERIDIAN = 8;
+    QUERY_CATEGORY_SOURCE = 9;
+    SERVER_EXPERIMENT = 10;
+    SHIELD_ID = 11;
+    SPAM = 12;
+    TEST_CODE = 13;
+    UNIFIED_USER_DATA_SOURCE = 14;
+    USER_ATTRIBUTE = 15;
+    USER_EXPERIMENT_ID_AND_TEST_CODE = 16;
  }
 
  enum StatusType {
      UNKNOWN = 0;
-     SUCCESS = 1;
-     DATA_NOT_INJECTED = 2;
+     DATA_NOT_INJECTED = 1;
+     SUCCESS = 2;
  }
 
  enum Pipeline {
     UNKNOWN_PIPELINE = 0;
-    HOLOGRAM_DEFAULT = 1;
-    HOLOGRAM_SPAM = 2;
-    HOLOGRAM_USER_ATTRIBUTE = 3;
-    HOLOGRAM_USER_EXPERIMENT_ID_AND_TEST_CODE = 4;
-    HOLOGRAM_ANDROID_SPAM_ABUSE = 5;
-    HOLOGRAM_PRIME_MERIDIAN = 6;
+    HOLOGRAM_ANDROID_SPAM_ABUSE = 1;
+    HOLOGRAM_DEFAULT = 2;
+    HOLOGRAM_PRIME_MERIDIAN = 3;
+    HOLOGRAM_SPAM = 4;
+    HOLOGRAM_USER_ATTRIBUTE = 5;
+    HOLOGRAM_USER_ATTRIBUTE = 6;
 }
