+++ hologram_dashboard/fetcher/proto/hologram_availability.proto
