+++ hologram_dashboard/fetcher/proto/hologram_config.proto
